+++ src/threadpool.hpp
@@ -2,6 +2,6 @@
 #define THREADPOOL_H
 extern unsigned long numberOfRepairScripts;
 bool ThreadPoolAddTask(void *(*)(void*), void *, bool);
-bool ThreadPoolNew(int threads = numberOfRepairScripts);
+bool ThreadPoolNew(size_t threads = numberOfRepairScripts);
 bool ThreadPoolCancel(void);
 #endif
